+++ ItunesMedia/Library/Extensions/UICollectionView+default.swift
@@ -1,0 +1,26 @@
+//
+//  UICollectionView+default.swift
+//  ItunesMedia
+//
+//  Created by Gustavo Lopez on 12/27/19.
+//  Copyright © 2019 Gustavo Lopez. All rights reserved.
+//
+
+import UIKit
+
+extension UICollectionView {
+    static func defaultCollectionView() -> UICollectionView {
+        let layout = UICollectionViewFlowLayout()
+        layout.scrollDirection = .vertical
+        
+        let collectionView = UICollectionView(
+            frame: .zero,
+            collectionViewLayout: layout
+        )
+        collectionView.backgroundColor = .white
+        collectionView.showsVerticalScrollIndicator = false
+        collectionView.translatesAutoresizingMaskIntoConstraints = false
+        return collectionView
+    }       
+}
+