+++ ItunesMedia/Resources/R.swift
@@ -19,6 +19,7 @@
 // swiftlint:disable identifier_name line_length nesting type_body_length type_name
 internal enum R {
   internal enum Base {
+    internal static let available = ImageAsset(name: "available")
     internal static let placeholder = ImageAsset(name: "placeholder")
     internal static let results = ImageAsset(name: "results")
   }